requirements.txt
streamlit
qdrant-client
sentence-transformers
rank_bm25
transformers
torch
tika
pytesseract
openai-whisper
spacy
networkx
scikit-learn
pymongo
requests
python-dotenv
Pillow
pymupdf
python-pptx